Flow analysis: fix parameter type of `ifStatement_thenBegin`.
For a while during the development of flow analysis for patterns, I
was using `ifStatement_thenBegin` for both ordinary if-statements and
if-case statements, distinguishing the two by having the caller pass a
`null` value for `condition` in the latter case. Later I reworked the
API, adding `ifCaseStatement_begin`,
`ifCaseStatement_afterExpression`, and `ifCaseStatement_thenBegin` to
handle if-case statements. But I forgot to restore the old signature
for `ifStatement_thenBegin`.

This change restores the old signature, in which the `condition`
argument is non-nullable. There should be no functional change, since
all callers were passing non-null values already.

Change-Id: I8c275f0a5be453c7de6052262a8a87eaac8d75ee
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/306300
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>