Flow analysis: clean up type parameters.
The `Expression` type parameter of `FlowAnalysis` is always a subtype
of `Node` in practice; this change updates the bounds on
`FlowAnalysis`, `FlowAnalysisDebug`, `_FlowAnalysisImpl`, and
`_LegacyTypePromotion` to reflect that.

Also, this change updates the redirecting factory constructor
`FlowAnalysis.legacy` so that it properly passes all the type
parameters over to `_LegacyTypePromotion`. There's no functional
change here, because type inference infers the correct parameters in
practice, but I prefer not to rely on that for correctness.

Change-Id: Ieacbe11dbb859557f27b5e2937b29381e8e934b1
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/306315
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>