[vm/aot/tfa] Fix building of data flow summary in case of break from try/finally
In case of a non-local jump (via 'break' or 'continue') which crosses
try/finally block, 'finally' block is executed, so outgoing state of
the 'finally' block should be used at the target of the jump,
instead of the state at break/continue.

In addition to fixing this bug, this change also slightly improves
precision when merging states in try/catch and try/finally based on
the conditions.

TEST=pkg/vm/testcases/transformations/type_flow/summary_collector/control_flow.dart

Change-Id: Id0aa9fa43e73afe2b7eee217837c11dc586b6008
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/306663
Commit-Queue: Alexander Markov <alexmarkov@google.com>
Reviewed-by: Slava Egorov <vegorov@google.com>