[analysis_server] Extract base server functionality from protocol-specific callbacks
This is just a refactor with no functional changes.

To simplify future work of supporting both protocols against a single server, this removes some common functionality out of each server into the base server (or in this case, some base callbacks that operate on the base server).

It's important when the server runs with both protocols that we don't accidentally duplicate any work, so stripping down the servers (and callbacks) to only the protocol-specific parts should make this easier.

Change-Id: I040368078b992d0c41bd8de10b09e38d3957f004
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/307081
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>