[dart2wasm] Fix initializing exception variables
Currently setting exception variables in `catch` blocks assumes that the
variables are not captured and always creates a new local for the
variable.

This CL adds a new function for initializing a variables. The
implementation is similar to `VariableDeclration` visitor and handles
captured (updated or not updated) and non-captured variables. This
function is then used in `catch` blocks to initialize exception and
stack trace variables.

New passing tests:

- co19/LibTest/async/Zone/handleUncaughtError_A01_t04
- lib/async/future_test

Fixes #52556.

Change-Id: I428d8f30ab1509b5b4439f8136190134741dd4ba
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/307042
Reviewed-by: Joshua Litt <joshualitt@google.com>
Reviewed-by: Aske Simon Christensen <askesc@google.com>
Commit-Queue: Ömer Ağacan <omersa@google.com>