[dds/dap] Respond to DAP launch request before sending runInTerminalRequest
Currently when the user configures a Dart CLI program to be run in the terminal, we ask VS Code to launch it in the terminal before we respond to the launch request. This (depending on settings) can cause VS Code to show the terminal and then switch to the Debug Console (which is both confusing and frustrating for users that opt-in to do this to access stdin of the terminal process).

This change reverses the order when using `runInTerminal` so that `launch` is responded to first.

See https://github.com/microsoft/vscode/issues/168295

Fixes https://github.com/Dart-Code/Dart-Code/issues/4287

Change-Id: If9a221361cc3329fb86fb17ba532043266b3438c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/307481
Reviewed-by: Ben Konyi <bkonyi@google.com>
Commit-Queue: Ben Konyi <bkonyi@google.com>