[js_runtime] More efficient GeneralConstantMap
Using a custom LinkedHashMap for the backing store was not efficient
for Strings and numbers since it bypassed the efficient storage for
those types and forced hash buckets to be used.

Adding a new JsConstantLinkedHashMap targeting primitive equality allows the efficient path to be used.

JsConstantLinkedHashMap is very similar to the implementation of LinkedHashMap.identity, so this class has been renamed moved to be next to JsConstantLinkedHashMap.


Change-Id: Iff97924dd645071d3ed767ce41c2dc5a41c779a2
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/309640
Reviewed-by: Mayank Patke <fishythefish@google.com>
Commit-Queue: Stephen Adams <sra@google.com>